Fix handling of default values for scanners (fixes #60)
The default values for dropdowns and checkboxes were not being
populated, and then the client side was initializing them to blank
strings, which caused the default values to be wrong for script and
pingless fields. This commit initializes default values for these field
types, removes the int() coercion for pingless, and changes how the
client side intializes values from default (since "false" is a valid
default).